System: Interactables:
Created -
- Base
- Keys
- Doors
- Coins
- Scoring System

-- Base Interactable:
Nothing of importance was really added, was simply a way to create child classes for scalable and flexible functionality

-- Keys
KeyID Enum was created to keep track of KeyID's. When the player collides with the Key BP, that key value is stored into a "Keychain" Array which will be checked by the door

-- Door
Door uses a simple timeline to open after checking if the player (on interact) as the correct KeyID (instance editable) in its Keychain Array

-- Coins
Created Coin Enum and in constructio scipt, gave numeric values to Enum's Low, Med, and High. Coins are set to rotate and portray correct coin values on surface. When coins are interacted with, coins access the Actor Component attached to the Player, "Scoring System" where CollectCoin() is called

-- Scoring System
Scoring system was initially placed as a BP but it needed a physical presence in the level. Instead, I turned it into an Actor Component so its always present as long as the player is present in the level as well